requirements.txt
gsplat
numpy
pycolmap
viser
opencv-python
tqdm
pytorch_msssim
scikit-learn
plyfile
torchmetrics
transformers
websockets
ffmpeg-python